requirements.txt
discord.py
python-dotenv
fastapi[all]
pandas
openpyxl
psutil
psycopg2